Fixed https://github.com/oracle/helidon/issues/299 - java.lang.IllegalMonitorStateException: attempt to unlock read lock, not locked by current thread
