+++ config/config/src/test/java/io/helidon/config/ConfigLeafImplTest.java
@@ -41,7 +41,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 /**
  * Tests {@link Config} API in case the node is {@link Config.Type#VALUE} type, i.e. {@link ConfigLeafImpl}.
  */
-public class ConfigLeafImplTest extends AbstractConfigImplTest {
+public class ConfigLeafImplTest extends AbstractConfigImplTestBase {
 
     public static Stream<TestContext> initParams() {
         return Stream.of(
